feat: refactored code to reduce the use of scss files
