+++ src/presentation/components/register/candidate-register/index.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import Button from "@/presentation/components/button";
-import styles from './styles.module.scss'
+import styles from '../styles.module.scss'
 import Input from "@/presentation/components/input"
 
 import { zodResolver } from "@hookform/resolvers/zod";
@@ -58,7 +58,7 @@ export default function CandidateRegister({listOfUFs}: {listOfUFs: {name: string
         })()
       })
   return (
-      <form className={styles.register_candidate_form} onSubmit={handleSubmit(onSubmit)}>
+      <form className={styles.register_form} onSubmit={handleSubmit(onSubmit)}>
         <div className={styles.form_rows}>
           <Input text="Nome Completo" uppercase forName="name" type="text" register={register} name="name" maxLength={60} placeholder="Ex: João da Silva" />  
 
@@ -87,7 +87,9 @@ export default function CandidateRegister({listOfUFs}: {listOfUFs: {name: string
           <Input forName="confirm_password" text="Confirmar Senha" uppercase type="password" register={register} name="confirm_password" maxLength={100} placeholder="" />
         </div>
 
-        <Button size="large" text="REGISTRAR" type="submit" />
+        <div className={styles.button_div}>
+          <Button size="large" text="REGISTRAR" type="submit" />
+        </div>
       </form>
   );
 }
