+++ src/presentation/components/register/styles.module.scss
@@ -1,15 +1,20 @@
 @import '@/app/globals.scss';
 
-.register_candidate_form  {
+.register_form  {
     align-items: center;
     display: flex;
     flex-direction: column;
-    gap: 30px;
+    gap: 56px;
 }
 
 .form_rows  {
     display: flex;
     flex-direction: row;
-    gap: 72px;
+    flex-wrap: wrap;
+    gap: 56px;
     width: 100%;
+}
+
+.button_div {
+    margin-top: 40px;
 }
