+++ src/presentation/components/select/styles.module.scss
@@ -26,9 +26,4 @@
     height: 50px;
     padding: 10px;
     outline: none;
-
-    @media  (max-width: 1440px) {
-        height: 50px;
-        width: 530px;
-    }
 }
